backend/whisper-python-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
openai>=1.0
openai-whisper>=20231117
